+++ Tests/Runtime/NativeStackTraceParser.cs.meta
